requirements.txt
py
pytest==3.0.2
requests==2.7.0
six==1.9.0
wheel==0.24.0
python-dotenv
pyTelegramBotAPI
flask
pip==19.2.3